requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cssselect2==0.4.1
dataclasses==0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.0.5
django-debug-toolbar==3.2
django-solo==1.1.5
django-widget-tweaks==1.4.8
git+https://github.com/harizMunawar/djongo@e8e8f6dab3657086b2b665fae8f5e6f27786a336#egg=djongo
dnspython==2.0.0
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jdcal==1.4.1
lockfile==0.12.2
msgpack==0.6.2
numpy==1.20.1
odfpy==1.4.1
openpyxl==3.0.6
packaging==20.3
pandas==1.2.2
pep517==0.8.2
Pillow==8.1.0
progress==1.5
pycparser==2.20
pymongo==3.7.2
pyparsing==2.4.6
Pyphen==0.10.0
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.5
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.2.4
tinycss2==1.1.0
urllib3==1.25.8
WeasyPrint==52.2
webencodings==0.5.1
whitenoise==5.2.0
